pyproject.toml
[tool.poetry]
name = "mathefragen-de"
version = "0.1.0"
description = ""
authors = [
    "Christian Meter <christian@schnaq.com>",
    "Timo Stark <timo.stark@myses.de>"
]
readme = "README.md"
packages = [{ include = "mathefragen" }]

[tool.poetry.dependencies]
python = "^3.12"
Markdown = "^2.6.11"
boto3 = "^1.34.147"
cloudinary = "^1.40.0"
cryptography = "^43.0.0"
django-admin-sortable2 = "^2.2.1"
django-appconf = "^1.0.6"
django-compressor = "^4.5.1"
django-environ = "^0.11.2"
django-extensions = "^3.2.3"
django-hashers-passlib = { git = "https://github.com/mathiasertl/django-hashers-passlib.git" }
django-ipware = "^7.0.1"
django-markdownify = "^0.9.5"
django-redis-sessions = "^0.6.2"
django-storages = "^1.14.4"
django-user-agents = "^0.4.0"
django = "^5.0.7"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework = "^3.15.1"
drf-spectacular = "^0.27.2"
drf-spectacular-sidecar = "^2024.7.1"
google-cloud-vision = "^3.7.3"
html2markdown = "^0.1.7"
ordereddict = "^1.1"
paypal-checkout-serversdk = "^1.0.3"
piexif = "^1.1.3"
Pillow = "^10.3.0"
psycopg2-binary = "^2.9.9"
pyfcm = "^1.5.4"
pytz = "^2024.1"
rcssmin = "^1.1.2"
requests = "^2.32.3"
uwsgi = "^2.0.26"
websocket-client = "^1.8.0"
zipp = "^3.19.2"
gunicorn = "^22.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

